Add files via upload
Wordle project made in school. This is a basic version of the game that lets you pick the word-length and selects a word from a limited pool of them. Then you guess the word within the terminal, with different symbols representing the letters you got right and the ones that are also in the same position (the green and yellow letters).